backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
joblib==1.4.2
xgboost==2.0.3
shap==0.45.1
scikit-learn==1.4.2
imbalanced-learn==0.12.2
mlflow==2.12.2
pandas==2.2.2
numpy==1.26.4
pydantic==2.7.1
python-dotenv==1.0.1
matplotlib==3.8.4
